+++ dk.common/src/main/java/com/danieljhkim/dsearch/common/enums/HealthStatus.java
@@ -1,0 +1,5 @@
+package com.danieljhkim.dsearch.common.enums;
+
+public enum HealthStatus {
+	UP, DOWN, DEGRADED
+}