+++ dk.query-node/src/main/java/com/danieljhkim/dsearch/querynode/QueryNodeApplication.java
@@ -4,6 +4,8 @@ import java.io.IOException;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import com.danieljhkim.dsearch.common.config.AppConfig;
+import com.danieljhkim.dsearch.common.config.ConfigLoader;
 import com.danieljhkim.dsearch.common.grpc.NodeClientManager;
 import com.danieljhkim.dsearch.common.health.HealthHttpServer;
 import com.danieljhkim.dsearch.proto.cluster.NodeRole;
@@ -22,12 +24,13 @@ public class QueryNodeApplication {
 		int grpcPort = Integer.parseInt(System.getenv("QUERY_NODE_PORT"));
 		int healthPort = Integer.parseInt(System.getenv("QUERY_NODE_HEALTH_PORT"));
 
+		AppConfig appConfig = ConfigLoader.load();
 		NodeClientManager<IndexServiceGrpc.IndexServiceBlockingStub> nodeClientManager = NodeClientManager
 				.loadClientManager(NodeRole.NODE_ROLE_INDEX,
 						IndexServiceGrpc::newBlockingStub);
 		SearchExecutor searchExecutor = new SearchExecutor(nodeClientManager);
 		BaseIndexService indexService = new IndexService(nodeClientManager);
-		QueryNodeServer queryNodeServer = new QueryNodeServer(grpcPort, searchExecutor, indexService);
+		QueryNodeServer queryNodeServer = new QueryNodeServer(grpcPort, searchExecutor, indexService, appConfig);
 		HttpServer healthServer = HealthHttpServer.start(healthPort, "query-node");
 
 		Runtime.getRuntime().addShutdownHook(new Thread(() -> {